Move IsPluginConfig validator to validators file
